refactor: extract shared formatLogEntry/colorLevel to format.ts
Deduplicate identical log formatting functions that existed independently in
both dashboard.ts and logs.ts. The canonical version (from logs.ts, which
includes sessionId display) now lives in lib/format.ts.

Also fixes the --since option in `ao dashboard logs` which was declared
but silently ignored (imports were unused).
