+++ packages/cli/src/lib/format.ts
@@ -1,5 +1,5 @@
 import chalk from "chalk";
-import type { CIStatus, ReviewDecision, ActivityState } from "@composio/ao-core";
+import type { CIStatus, ReviewDecision, ActivityState, LogEntry } from "@composio/ao-core";
 
 export function header(title: string): string {
   const line = "─".repeat(76);
@@ -126,6 +126,25 @@ export function parseSinceArg(since: string): Date {
     unit === "h" ? value * 3_600_000 :
     value * 86_400_000;
   return new Date(Date.now() - ms);
+}
+
+/** Color-code log level for terminal display. */
+export function colorLevel(level: LogEntry["level"]): string {
+  switch (level) {
+    case "error": return chalk.red("ERR");
+    case "warn": return chalk.yellow("WRN");
+    case "stderr": return chalk.red("err");
+    case "stdout": return chalk.dim("out");
+    case "info": return chalk.blue("inf");
+  }
+}
+
+/** Format a log entry for terminal display. */
+export function formatLogEntry(entry: LogEntry): string {
+  const ts = new Date(entry.ts).toLocaleTimeString();
+  const level = colorLevel(entry.level);
+  const session = entry.sessionId ? chalk.cyan(entry.sessionId) + " " : "";
+  return `${chalk.dim(ts)} ${level} ${session}${entry.message}`;
 }
 
 // eslint-disable-next-line no-control-regex
